add editable timeout

+++ config.py
@@ -14,7 +14,7 @@ class Config(BaseSettings):
     API_SOAP_USER: str
     API_SOAP_PAS: str
     API_CALLS_DELAY: float = 0
-    API_CALLS_DELAY_TIMEOUT_ERROR: float = 0
+    API_CALLS_DELAY_TIMEOUT_ERROR: int = 0
     APP_DATA_DIR: str
  
     DB_URI: str
